Support for textured models.

+++ Exort/src/com/doobs/exort/entity/Entity.java
@@ -3,7 +3,7 @@ package com.doobs.exort.entity;
 import com.doobs.exort.level.*;
 
 public class Entity {
-	protected float x, y, z;
+	protected double x, y, z;
 
 	protected Level level;
 
@@ -11,19 +11,19 @@ public class Entity {
 		this(0, 0, 0, null);
 	}
 
-	public Entity(int x, int z) {
+	public Entity(double x, double z) {
 		this(x, 0, z, null);
 	}
 
-	public Entity(int x, int y, int z) {
+	public Entity(double x, double y, double z) {
 		this(x, y, z, null);
 	}
 
-	public Entity(int x, int z, Level level) {
+	public Entity(double x, double z, Level level) {
 		this(x, 0, z, level);
 	}
 
-	public Entity(int x, int y, int z, Level level) {
+	public Entity(double x, double y, double z, Level level) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
@@ -31,27 +31,27 @@ public class Entity {
 	}
 
 	// Getters and Setters
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public float getZ() {
+	public double getZ() {
 		return z;
 	}
 
-	public void setZ(float z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
 }
